Checkbox moved into CheckboxComponent
